DEBUG: Fix root ASN1 conversion after pyasn1 0.3.6 changes: key val
THIS IS A DEBUG-ONLY COMMIT: The debug content should be removed
in a rebase before it is merged!

For some reason, sometimes the prettyPrint() function on the pyasn1
object publicKeyValue prefixes the hex string produced with 0x and
sometimes it does not. Strip it when it's there.

As a precaution, the same check-and-strip was added to two other
conversions for root metadata, both for key id, even though I have
not observed cases where the pyasn1 prettyPrint did not add '0x'
to the start of the string....